+++ docs/source/changes.md
@@ -26,7 +26,8 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 - {pull}`498` fixes an error when using {class}`~pytask.Task` and
   {class}`~pytask.TaskWithoutPath` in task modules.
 - {pull}`500` refactors the dependencies for tests.
-- {pull}`508` catches objects that pretend to be a {class}`~pytask.PTask`.
+- {pull}`508` catches objects that pretend to be a {class}`~pytask.PTask`. Fixes
+  {issue}`507`.
 
 ## 0.4.2 - 2023-11-08